vpn-agent: Implement VPN agent setting to keep credentials
Implement a setting "KeepCredentials" into vpn-agent.c. This boolean
value can be used the registered VPN agent to detect if the request
itself has storage and retrieval forbidden but still does not wish to
remove the credentials.

A example case of this is the OpenVPN Encrypted Private Key password
that is to be kept in memory only. But by disabling storage and
retrieval for the Private Key password request should propagate into
clearing the credentials for the connection.