vpn-provider: Support split routing option for VPN providers
Support SplitRouting option in vpnd as well. The value is passed as
boolean, and read from the settings file as boolean. A warning is
printed with invalid error if attempted to be set as string. For this,
vpn_provider_set_boolean() is added.

In addition handling of default route for VPNs is improved with checks
for split routing and already set route. This is to ensure that if a
non-split routed VPN does not have default route set in the routes it
should be added in order to guarantee that traffic is routed over the
VPN. This check is done also in provider_append_routes() to track if the
default route is included in the provided routes.