provider: Support split routing in VPN properties
Split routing variable should be known also the VPNs and exposed in the
VPN properties. Add a function connman_provider_set_split_routing()
for plugins to be able to set the split routing value.

In to ensure the proper handling of split routing value changes utilize
the separated service moving __connman_service_move_service() function
to reorder the services. When a VPN is to be set as split routed it ha
to be used as a target, and when VPN is not split routed it should be
regarded as the default service.

If the VPN is not connected only the split routing value is changed for
the service. If there is an error in moving the connected service use
__connman_service_split_routing_changed() to signal vpnd about the
current status of split routing.

To ensure that setting file has the SplitRouting value if changed in VPN
settings by vpnd the vpn_service must be saved. This should be done in
provider.c since in service.c the saving is done explicitly as well
after changing SplitRouting value.

To allow checking the split routing status also in VPN plugin add
connman_provider_is_split_routing() to get the split routing value
from the actual service. By default returns false if the provider or
service is omitted.

Add getter for provider family: connman_provider_get_family(). Return
AF_UNSPEC with NULL provider.