inet: Refactor with getifaddrs() and add network route getter function
Create new internal get_interface_addresses() function that utilizes
simple structure to get the IP address, netmask, broadcast address
and/or P-t-P address using getifaddrs() in a single run. This allows to
define whether to accept unspecified addresses (checked with
is_addr_unspec()) and to require the IP address to be link local address
(checked with is_addr_ll()).

If the set requirements for unspec/link local are not met, -ENOENT is
returned as error and all the pointers in the struct interface_address
are NULL. With invalid address family, or missing struct -EINVAL is
returned. Or the error of getifaddrs() is returned if the call fails.

When get_interface_addressses() is used directly, the pointer array in
struct interface_address must have struct in_addr* for AF_INET and
struct in6_addr* for AF_INET6 set as family. The types defined in enum
ipaddr_type are to be used as a position for the different addresses
requested.

Cleaned up a lot of existing code to have only one use of getifaddrs()
via the get_interface_addresses() by just filling in the structure with
the required parameters.

Created new functions connman_inet_get_route_addresses() and
connman_inet_ipv6_get_route_addresses() for getting the string
representation of all the required address for a route. Parameter
destination (for P-t-P) is optional.