inet: Do not add broadcast address for P2P/VPNs
Adding broadcast address for VPNs is not permitted by "ip route" and it
messes up the use of getifaddrs() for getting the destination address.
Adding broadcast address for P2P connections is also unnecessary. This
change adds a simple boolean to toggle whether the address change
concerns a VPN or any other service, to toggle adding of IFA_BROADCAST.
Naturally this only affects IPv4 VPNs.

The issue fixed here is easy to reprocude by connecting a VPN, then
checking the destination address with ioctl() using SIOCGIFDSTADDR flag for
IFF_POINTOPOINT type interface, and then using getifaddrs() for the same
interface by checking the ifa_flags for flag IFF_POINTOPOINT to get the
destination address. Or simply use small test app for doing the check:
https://pastebin.com/1HMuXYHK

The toggle for P2P/VPN connections is saved in struct connman_ipaddress,
and is set at allocation (connman_ipaddress_alloc()). This setting is as
permanent as the IP address family set at allocation and thus, is not
cleared/changed when the addresses are changed in the struct.

Also changed connman_inet_clear_ipv6_address() to use the struct
connman_ipaddress as other set/clear_ipv4/6_address functions do.