ipconfig: Add forced disabling of IPv6 and support IPv6 method restore
Support complete disabling of IPv6 on both system and interface levels.
Disabling of IPv6 completely albeit temporarily is required when there
can be multiple connected technologies and a VPN is connected over IPv4
using an dual-stack transport and/or there is another technology with
dual-stack or only IPv6 in use. This approach effectively ensures that
no data from an IPv4 VPN leaks to any other interface. A DNS server
returning replies to AAAA requests may return an IPv6 address, which
could be then used over IPv6 routing on another technolology that is not
acting as the transport of the VPN.

Add getter and setter for the internal IPv6 status. When enabling check
the status with connman_inet_is_ipv6_supported(). It is not the
intention to disable IPv6 system-wide since there may be interfaces that
should not be managed by ConnMan and are handled for other OS needs
where IPv6 is required, e.g., VoLTE may be one of them.

Add a "ipv6_force_disabled" toggle to indicate that the particular
device tied to ipconfig should not have IPv6 enabled. This toggle is to
be used in scenarios where IPv6 should be prevented from reinstating
unless forcefully enabled. Added a separate setter function for the
value (__connman_ipconfig_set_force_disabled_ipv6()).

When checking IPv6 enabled state return false if ipv6_force_disabled is
set as that is the real status of IPv6. The /proc disable_ipv6 can
change when kernel processes ICMPv6 (RA/RS) packets and brings the IPv6
interface up so using only the disable_ipv6 value is not consistent
enough in case when IPv6 has been forcefully disabled.

This change will prevent from changing the IPv6 status until it has
it has been forcefully enabled, until that -EOPNOSUPP is returned.
The most prominent use case for this is to prevent data as well as DNS
leak to IPv6 when IPv4 only VPN is connected over a transport supporting
both IPv4 and IPv6 connectivity.

In addition to changing disable_ipv6 value also the autoconf option is
managed. This is done in order to control the address setup for the
interface. It is worthy to note that disable_ipv6 may be temporarily
set to 0 by the kernel. This can happen when LL address is temporarily
set as a typical behavior of IPv6 on some configurations to handle
ICMPv6. After a while the value of disable_ipv6 as well as LL address
are reset/removed by kernel.

Add functions to save and restore old IPv6 method. This is useful when
IPv6 is disabled termporarily for the time IPv4 VPN is used. First save
the IPv6 method before disabling and call restore when IPv6 needs to be
set up again using the old method.