network: Adopt to new IPv6 disabled functionality of ipconfig.c
If a new connection comes up/is manually connected when IPv6 is
internally disabled for a IPv4 VPN the IPv6 configuration for the
network must be prevented. If the different IPv6 method functions do
this then an error is received from ipconfig.c which will propagate to
disabling of the new connection completely as network error is
triggered. This way IPv6 is silently ignored and forcefully set as
disabled to be re-enabled when VPN is disconnected.

Disable force disabled in IPv6 ipconfig when clearing the IPv6 address
in set_disconnect(). This is required when re-enabling IPv6 for kernel
to clear the values in order to return the proper /proc values for the
interface in case IPv6 was disabled for IPv4 VPN.

Adapt to ipconfig.c changes by reporting back potential errors.