provider: Toggle IPv6 on the transport of IPv4 VPN connection
Add support to disable/enable IPv6 on the transport of the VPN that uses
IPv4. This change eliminates the data and DNS leak to IPv6 when
dual-stack transport is used on a IPv4 only VPN. Otherwise with an AAAA
record for a requested hostname the traffic can bypass the VPN to
transport's IPv6 network if the DNS server of the VPN serves both A
and AAAA requests.

If multiple connection technologies (SingleConnectedTechnology omitted
or false) are in use IPv6 support is changed internally as well. The
value of SingleConnectedTechnology does not change run-time so there is
no possibility for inconsistent state. When re-enabling IPv6 enable the
internal IPv6 suppor prior to resuming IPv6 functionality on service
basis. When disabling IPv6 do the internal disabling after the services
are handled. This is due to the changes that were required for network.c
to get new connections enabled but IPv6 disabled if IPv6 is internally
disabled.

Disable IPv6 when state changes to READY (also ONLINE but that is never
used with VPNs) for IPv4 provider. Allow to change IPv6 only for VPNs
that are connected (READY).

When provider state changes to DISCONNECT or FAILURE re-enable IPv6 the
transport using the recorded method. Prevent running the function as a
loop within loop. This might happen when current transport service is
disconnected that in turn causes the VPN to be disconnected as well.
Then while looping through the services another disconnect on the
transport will get provider_indicate_state() called again.