WIP: wireguard: Fix FQDN by using the resolved IP as the gateway
WIP: [wireguard] Fix FQDN by using the resolved IP as the gateway. JB#63273

This fixes the hostname use with WireGuard by using the resolved IP as
the gateway for the ipaddress. This address will be sent to connmand for
setting up the routes and DNS will be otherwise broken if the FQDN is
set "as is" as the gateway.

By returning also the resolved IP in string format in the initial
resolve, and setting it as the gateway for the address for the interface
in in the ipaddress.c the FQDN using WireGuard services can be used.

TODO: make this affect to the possible change of the endpoint after DNS
reresolve to change the routes and IP address with the same approach.